requirements.txt
pyusb
click>7
